Split up IndexHint
